refactor(client): rename account wallet module
