+++ packages/client/src/index.ts
@@ -7,7 +7,6 @@ export type * from '@polymarket/bindings/gamma';
 export { WalletType } from '@polymarket/bindings/gamma';
 export type * from '@polymarket/bindings/relayer';
 export * from './abis';
-export type { AccountIdentity } from './account';
 export type {
   RelayerApiKeyConfig,
   RemoteBuilderSigningConfig,
@@ -26,3 +25,4 @@ export * from './errors';
 export * from './hmac';
 export type * from './pagination';
 export * from './types';
+export type { AccountIdentity } from './wallet';
